- ADDED: is_multi() to MY_Array_helper - DATAMODEL: Creating config had a bug - DATAMODEL: supports foreignkeys with underscore at end (for multiple foreign keys to same foreign table)
git-svn-id: file:///Users/jan/Dropbox/svn/FlexyAdmin/trunk@3595 0c02b2c5-a8bd-4e35-85b2-cf8cacb96951

Former-commit-id: f3723a6d1ec85d3d9d781d2cfe81e67ea17c3254
Former-commit-id: 93f59ec36edafb2aa6d3e053a36fa6a22c841a5a
Former-commit-id: 073e85886793a8c498f2fc34186cbd5aec2a87e7 [formerly 20618c7689d8e50608f155b7f3622c33d1874c23]
Former-commit-id: 46ccada19b093b83782dc1089ca669fcba793c29

+++ sys/flexyadmin/libraries/plugins/Plugin_move_site.php
@@ -43,13 +43,13 @@ class Plugin_move_site extends Plugin {
     }
     
     
-    $this->add_message('<h2>Merge old database with fresh database</h2>');
-    $old_db = $this->config['db'];
-    if ($old_db) {
-      $this->oldDB = $this->CI->load->database( $this->config['db'], TRUE);
-      $this->truncate_demo_tables();
-      $this->import_database();
-    }
+    // $this->add_message('<h2>Merge old database with fresh database</h2>');
+    // $old_db = $this->config['db'];
+    // if ($old_db) {
+    //   $this->oldDB = $this->CI->load->database( $this->config['db'], TRUE);
+    //   $this->truncate_demo_tables();
+    //   $this->import_database();
+    // }
 	
   	return $this->view('admin/plugins/plugin');
   }
@@ -64,8 +64,10 @@ class Plugin_move_site extends Plugin {
   private function truncate_demo_tables() {
     $ul=array();
     foreach ($this->config['truncate_demo_tables'] as $table) {
-      $this->CI->db->truncate( $table );
-      $ul[]=$table;
+      if ($this->CI->db->table_exists($table)) {
+        $this->CI->db->truncate( $table );
+        $ul[]=$table;
+      }
     }
     $this->add_message('<h3>Truncated demo tables:</h3>');
     $this->add_message(ul($ul));
